Changed how files are created on the destination system to create their requested directory tree if it is not already present.
Added support for sending directories. The root name of the folder will be the final directory name passed in the destination string but otherwise the folder structure will be preserved.
Ie. sending "C:\Test" to the destination as "C:\NewFolder" will place all of the contents of Test into a folder named NewFolder.